task2/requirements.txt
aiosqlite==0.20.0
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
packaging==24.0
passlib==1.7.4
pluggy==1.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pytest==8.1.1
pytest-asyncio==0.23.6
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
